+++ src/Recyclarr.Cli/Console/Settings/IDeleteCustomFormatSettings.cs
@@ -1,0 +1,10 @@
+namespace Recyclarr.Cli.Console.Settings;
+
+public interface IDeleteCustomFormatSettings
+{
+    public string InstanceName { get; }
+    public IReadOnlyCollection<string> CustomFormatNames { get; }
+    public bool All { get; }
+    public bool Force { get; }
+    bool Preview { get; }
+}