+++ src/Recyclarr.Cli/Pipelines/CustomFormat/Api/ICustomFormatService.cs
@@ -8,5 +8,9 @@ public interface ICustomFormatService
     Task<IList<CustomFormatData>> GetCustomFormats(IServiceConfiguration config);
     Task<CustomFormatData?> CreateCustomFormat(IServiceConfiguration config, CustomFormatData cf);
     Task UpdateCustomFormat(IServiceConfiguration config, CustomFormatData cf);
-    Task DeleteCustomFormat(IServiceConfiguration config, int customFormatId);
+
+    Task DeleteCustomFormat(
+        IServiceConfiguration config,
+        int customFormatId,
+        CancellationToken cancellationToken = default);
 }