+++ src/Recyclarr.Cli/Processors/Delete/IDeleteCustomFormatsProcessor.cs
@@ -1,0 +1,8 @@
+using Recyclarr.Cli.Console.Settings;
+
+namespace Recyclarr.Cli.Processors.Delete;
+
+public interface IDeleteCustomFormatsProcessor
+{
+    Task Process(IDeleteCustomFormatSettings settings);
+}