+++ src/Recyclarr.Cli/Processors/ServiceProcessorsAutofacModule.cs
@@ -1,6 +1,7 @@
 using Autofac;
 using Autofac.Extras.Ordering;
 using Recyclarr.Cli.Processors.Config;
+using Recyclarr.Cli.Processors.Delete;
 using Recyclarr.Cli.Processors.Sync;
 
 namespace Recyclarr.Cli.Processors;
@@ -11,6 +12,8 @@ public class ServiceProcessorsAutofacModule : Module
     {
         base.Load(builder);
 
+        builder.RegisterType<ConsoleExceptionHandler>();
+
         // Sync
         builder.RegisterType<SyncProcessor>().As<ISyncProcessor>();
         builder.RegisterType<SyncPipelineExecutor>();
@@ -19,6 +22,9 @@ public class ServiceProcessorsAutofacModule : Module
         builder.RegisterType<ConfigManipulator>().As<IConfigManipulator>();
         builder.RegisterType<ConfigCreationProcessor>().As<IConfigCreationProcessor>();
         builder.RegisterType<ConfigListProcessor>();
+
+        // Delete
+        builder.RegisterType<DeleteCustomFormatsProcessor>().As<IDeleteCustomFormatsProcessor>();
 
         builder.RegisterTypes(
                 typeof(TemplateConfigCreator),