+++ src/Recyclarr.TrashLib/Config/IConfigurationRegistry.cs
@@ -1,0 +1,8 @@
+using Recyclarr.TrashLib.Config.Services;
+
+namespace Recyclarr.TrashLib.Config;
+
+public interface IConfigurationRegistry
+{
+    IReadOnlyCollection<IServiceConfiguration> FindAndLoadConfigs(ConfigFilterCriteria? filterCriteria = null);
+}