+++ src/Recyclarr.TrashLib/Config/Parsing/IConfigurationFinder.cs
@@ -4,5 +4,5 @@ namespace Recyclarr.TrashLib.Config.Parsing;
 
 public interface IConfigurationFinder
 {
-    IReadOnlyCollection<IFileInfo> GetConfigFiles(IReadOnlyCollection<IFileInfo>? configs = null);
+    IReadOnlyCollection<IFileInfo> GetConfigFiles();
 }