+++ src/Recyclarr.TrashLib/ExceptionTypes/CommandException.cs
@@ -1,0 +1,9 @@
+namespace Recyclarr.TrashLib.ExceptionTypes;
+
+public class CommandException : Exception
+{
+    public CommandException(string? message)
+        : base(message)
+    {
+    }
+}